Move TRN request endpoints into VNext
Also moves tests into a child namespace by version.
