+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next
 
+### TRN Requests
+
+Added `POST /v3/trn-requests` and `GET /v3/trn-requests` endpoints.
+
 
 ## 20240101
