+++ TeachingRecordSystem/tests/TeachingRecordSystem.Api.Tests/V3/V20240101/CreateDateOfBirthChangeTests.cs
@@ -3,27 +3,24 @@ using JustEat.HttpClientInterception;
 using Microsoft.Xrm.Sdk;
 using Microsoft.Xrm.Sdk.Query;
 
-namespace TeachingRecordSystem.Api.Tests.V3;
+namespace TeachingRecordSystem.Api.Tests.V3.V20240101;
 
 [Collection(nameof(DisableParallelization))]  // Configures EvidenceFilesHttpClient
-public class CreateNameChangeTests : TestBase
+public class CreateDateOfBirthChangeTests : TestBase
 {
-    public CreateNameChangeTests(HostFixture hostFixture)
+    public CreateDateOfBirthChangeTests(HostFixture hostFixture)
         : base(hostFixture)
     {
         SetCurrentApiClient(new[] { ApiRoles.UpdatePerson });
     }
 
-    [Theory, RoleNamesData(except: [ApiRoles.GetPerson, ApiRoles.UpdatePerson])]
-    public async Task PostNameChanges_ClientDoesNotHavePermission_ReturnsForbidden(string[] roles)
+    [Theory, RoleNamesData(except: [ApiRoles.UpdatePerson])]
+    public async Task PostCreateDateOfBirthChange_ClientDoesNotHavePermission_ReturnsForbidden(string[] roles)
     {
         // Arrange
         SetCurrentApiClient(roles);
         var trn = await TestData.GenerateTrn();
-        var newFirstName = TestData.GenerateFirstName();
-        var newMiddleName = TestData.GenerateMiddleName();
-        var newLastName = TestData.GenerateLastName();
-
+        var newDateOfBirth = TestData.GenerateDateOfBirth();
         var evidenceFileName = "evidence.txt";
         var evidenceFileUrl = Faker.Internet.SecureUrl();
         var evidenceFileContent = Encoding.UTF8.GetBytes("Test file");
@@ -45,14 +42,12 @@ public class CreateNameChangeTests : TestBase
                 .RegisterWith(options);
         });
 
-        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/name-changes")
+        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/date-of-birth-changes")
         {
             Content = CreateJsonContent(new
             {
                 trn,
-                firstName = newFirstName,
-                middleName = newMiddleName,
-                lastName = newLastName,
+                dateOfBirth = newDateOfBirth,
                 evidenceFileName,
                 evidenceFileUrl
             })
@@ -66,30 +61,28 @@ public class CreateNameChangeTests : TestBase
     }
 
     [Theory]
-    [InlineData(false, "First", "Middle", "Last", "evidence.jpg", "https://place.com/evidence.jpg")]
-    [InlineData(true, null, "Middle", "Last", "evidence.jpg", "https://place.com/evidence.jpg")]
-    [InlineData(true, "First", "Middle", null, "evidence.jpg", "https://place.com/evidence.jpg")]
-    [InlineData(true, "First", "Middle", "Last", null, "https://place.com/evidence.jpg")]
-    [InlineData(true, "First", "Middle", "Last", "evidence.jpg", null)]
+    [InlineData(false, "1990-07-01", "evidence.jpg", "https://place.com/evidence.jpg")]
+    [InlineData(true, null, "evidence.jpg", "https://place.com/evidence.jpg")]
+    [InlineData(true, "1990-07-01", "evidence.jpg", "https://place.com/evidence.jpg")]
+    [InlineData(true, "1990-07-01", null, "https://place.com/evidence.jpg")]
+    [InlineData(true, "1990-07-01", "evidence.jpg", null)]
     public async Task Post_InvalidRequest_ReturnsBadRequest(
         bool includeTrn,
-        string? newFirstName,
-        string? newMiddleName,
-        string? newLastName,
+        string? newDateOfBirthString,
         string? evidenceFileName,
         string? evidenceFileUrl)
     {
         // Arrange
         var createPersonResult = await TestData.CreatePerson();
 
-        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/name-changes")
+        var newDateOfBirth = newDateOfBirthString is not null ? DateOnly.ParseExact(newDateOfBirthString, "yyyy-MM-dd") : (DateOnly?)null;
+
+        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/date-of-birth-changes")
         {
             Content = CreateJsonContent(new
             {
                 trn = includeTrn ? createPersonResult.Trn : "",
-                firstName = newFirstName,
-                middleName = newMiddleName,
-                lastName = newLastName,
+                dateOfBirth = newDateOfBirth,
                 evidenceFileName,
                 evidenceFileUrl
             })
@@ -107,9 +100,7 @@ public class CreateNameChangeTests : TestBase
     {
         // Arrange
         var trn = await TestData.GenerateTrn();
-        var newFirstName = TestData.GenerateFirstName();
-        var newMiddleName = TestData.GenerateMiddleName();
-        var newLastName = TestData.GenerateLastName();
+        var newDateOfBirth = TestData.GenerateDateOfBirth();
 
         var evidenceFileName = "evidence.txt";
         var evidenceFileUrl = Faker.Internet.SecureUrl();
@@ -132,14 +123,12 @@ public class CreateNameChangeTests : TestBase
                 .RegisterWith(options);
         });
 
-        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/name-changes")
+        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/date-of-birth-changes")
         {
             Content = CreateJsonContent(new
             {
                 trn,
-                firstName = newFirstName,
-                middleName = newMiddleName,
-                lastName = newLastName,
+                dateOfBirth = newDateOfBirth,
                 evidenceFileName,
                 evidenceFileUrl
             })
@@ -157,21 +146,17 @@ public class CreateNameChangeTests : TestBase
     {
         // Arrange
         var createPersonResult = await TestData.CreatePerson();
-        var newFirstName = TestData.GenerateFirstName();
-        var newMiddleName = TestData.GenerateMiddleName();
-        var newLastName = TestData.GenerateLastName();
+        var newDateOfBirth = TestData.GenerateChangedDateOfBirth(currentDateOfBirth: createPersonResult.DateOfBirth);
 
         var evidenceFileName = "evidence.txt";
         var evidenceFileUrl = Faker.Internet.SecureUrl();
 
-        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/name-changes")
+        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/date-of-birth-changes")
         {
             Content = CreateJsonContent(new
             {
                 trn = createPersonResult.Trn,
-                firstName = newFirstName,
-                middleName = newMiddleName,
-                lastName = newLastName,
+                dateOfBirth = newDateOfBirth,
                 evidenceFileName,
                 evidenceFileUrl
             })
@@ -189,9 +174,7 @@ public class CreateNameChangeTests : TestBase
     {
         // Arrange
         var createPersonResult = await TestData.CreatePerson();
-        var newFirstName = TestData.GenerateFirstName();
-        var newMiddleName = TestData.GenerateMiddleName();
-        var newLastName = TestData.GenerateLastName();
+        var newDateOfBirth = TestData.GenerateChangedDateOfBirth(currentDateOfBirth: createPersonResult.DateOfBirth);
 
         var evidenceFileName = "evidence.txt";
         var evidenceFileUrl = Faker.Internet.SecureUrl();
@@ -214,14 +197,12 @@ public class CreateNameChangeTests : TestBase
                 .RegisterWith(options);
         });
 
-        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/name-changes")
+        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/date-of-birth-changes")
         {
             Content = CreateJsonContent(new
             {
                 trn = createPersonResult.Trn,
-                firstName = newFirstName,
-                middleName = newMiddleName,
-                lastName = newLastName,
+                dateOfBirth = newDateOfBirth,
                 evidenceFileName,
                 evidenceFileUrl
             })
