+++ TeachingRecordSystem/tests/TeachingRecordSystem.Api.Tests/V3/V20240101/CreateNameChangeTests.cs
@@ -3,24 +3,27 @@ using JustEat.HttpClientInterception;
 using Microsoft.Xrm.Sdk;
 using Microsoft.Xrm.Sdk.Query;
 
-namespace TeachingRecordSystem.Api.Tests.V3;
+namespace TeachingRecordSystem.Api.Tests.V3.V20240101;
 
 [Collection(nameof(DisableParallelization))]  // Configures EvidenceFilesHttpClient
-public class CreateDateOfBirthChangeTests : TestBase
+public class CreateNameChangeTests : TestBase
 {
-    public CreateDateOfBirthChangeTests(HostFixture hostFixture)
+    public CreateNameChangeTests(HostFixture hostFixture)
         : base(hostFixture)
     {
         SetCurrentApiClient(new[] { ApiRoles.UpdatePerson });
     }
 
-    [Theory, RoleNamesData(except: [ApiRoles.UpdatePerson])]
-    public async Task PostCreateDateOfBirthChange_ClientDoesNotHavePermission_ReturnsForbidden(string[] roles)
+    [Theory, RoleNamesData(except: [ApiRoles.GetPerson, ApiRoles.UpdatePerson])]
+    public async Task PostNameChanges_ClientDoesNotHavePermission_ReturnsForbidden(string[] roles)
     {
         // Arrange
         SetCurrentApiClient(roles);
         var trn = await TestData.GenerateTrn();
-        var newDateOfBirth = TestData.GenerateDateOfBirth();
+        var newFirstName = TestData.GenerateFirstName();
+        var newMiddleName = TestData.GenerateMiddleName();
+        var newLastName = TestData.GenerateLastName();
+
         var evidenceFileName = "evidence.txt";
         var evidenceFileUrl = Faker.Internet.SecureUrl();
         var evidenceFileContent = Encoding.UTF8.GetBytes("Test file");
@@ -42,12 +45,14 @@ public class CreateDateOfBirthChangeTests : TestBase
                 .RegisterWith(options);
         });
 
-        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/date-of-birth-changes")
+        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/name-changes")
         {
             Content = CreateJsonContent(new
             {
                 trn,
-                dateOfBirth = newDateOfBirth,
+                firstName = newFirstName,
+                middleName = newMiddleName,
+                lastName = newLastName,
                 evidenceFileName,
                 evidenceFileUrl
             })
@@ -61,28 +66,30 @@ public class CreateDateOfBirthChangeTests : TestBase
     }
 
     [Theory]
-    [InlineData(false, "1990-07-01", "evidence.jpg", "https://place.com/evidence.jpg")]
-    [InlineData(true, null, "evidence.jpg", "https://place.com/evidence.jpg")]
-    [InlineData(true, "1990-07-01", "evidence.jpg", "https://place.com/evidence.jpg")]
-    [InlineData(true, "1990-07-01", null, "https://place.com/evidence.jpg")]
-    [InlineData(true, "1990-07-01", "evidence.jpg", null)]
+    [InlineData(false, "First", "Middle", "Last", "evidence.jpg", "https://place.com/evidence.jpg")]
+    [InlineData(true, null, "Middle", "Last", "evidence.jpg", "https://place.com/evidence.jpg")]
+    [InlineData(true, "First", "Middle", null, "evidence.jpg", "https://place.com/evidence.jpg")]
+    [InlineData(true, "First", "Middle", "Last", null, "https://place.com/evidence.jpg")]
+    [InlineData(true, "First", "Middle", "Last", "evidence.jpg", null)]
     public async Task Post_InvalidRequest_ReturnsBadRequest(
         bool includeTrn,
-        string? newDateOfBirthString,
+        string? newFirstName,
+        string? newMiddleName,
+        string? newLastName,
         string? evidenceFileName,
         string? evidenceFileUrl)
     {
         // Arrange
         var createPersonResult = await TestData.CreatePerson();
 
-        var newDateOfBirth = newDateOfBirthString is not null ? DateOnly.ParseExact(newDateOfBirthString, "yyyy-MM-dd") : (DateOnly?)null;
-
-        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/date-of-birth-changes")
+        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/name-changes")
         {
             Content = CreateJsonContent(new
             {
                 trn = includeTrn ? createPersonResult.Trn : "",
-                dateOfBirth = newDateOfBirth,
+                firstName = newFirstName,
+                middleName = newMiddleName,
+                lastName = newLastName,
                 evidenceFileName,
                 evidenceFileUrl
             })
@@ -100,7 +107,9 @@ public class CreateDateOfBirthChangeTests : TestBase
     {
         // Arrange
         var trn = await TestData.GenerateTrn();
-        var newDateOfBirth = TestData.GenerateDateOfBirth();
+        var newFirstName = TestData.GenerateFirstName();
+        var newMiddleName = TestData.GenerateMiddleName();
+        var newLastName = TestData.GenerateLastName();
 
         var evidenceFileName = "evidence.txt";
         var evidenceFileUrl = Faker.Internet.SecureUrl();
@@ -123,12 +132,14 @@ public class CreateDateOfBirthChangeTests : TestBase
                 .RegisterWith(options);
         });
 
-        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/date-of-birth-changes")
+        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/name-changes")
         {
             Content = CreateJsonContent(new
             {
                 trn,
-                dateOfBirth = newDateOfBirth,
+                firstName = newFirstName,
+                middleName = newMiddleName,
+                lastName = newLastName,
                 evidenceFileName,
                 evidenceFileUrl
             })
@@ -146,17 +157,21 @@ public class CreateDateOfBirthChangeTests : TestBase
     {
         // Arrange
         var createPersonResult = await TestData.CreatePerson();
-        var newDateOfBirth = TestData.GenerateChangedDateOfBirth(currentDateOfBirth: createPersonResult.DateOfBirth);
+        var newFirstName = TestData.GenerateFirstName();
+        var newMiddleName = TestData.GenerateMiddleName();
+        var newLastName = TestData.GenerateLastName();
 
         var evidenceFileName = "evidence.txt";
         var evidenceFileUrl = Faker.Internet.SecureUrl();
 
-        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/date-of-birth-changes")
+        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/name-changes")
         {
             Content = CreateJsonContent(new
             {
                 trn = createPersonResult.Trn,
-                dateOfBirth = newDateOfBirth,
+                firstName = newFirstName,
+                middleName = newMiddleName,
+                lastName = newLastName,
                 evidenceFileName,
                 evidenceFileUrl
             })
@@ -174,7 +189,9 @@ public class CreateDateOfBirthChangeTests : TestBase
     {
         // Arrange
         var createPersonResult = await TestData.CreatePerson();
-        var newDateOfBirth = TestData.GenerateChangedDateOfBirth(currentDateOfBirth: createPersonResult.DateOfBirth);
+        var newFirstName = TestData.GenerateFirstName();
+        var newMiddleName = TestData.GenerateMiddleName();
+        var newLastName = TestData.GenerateLastName();
 
         var evidenceFileName = "evidence.txt";
         var evidenceFileUrl = Faker.Internet.SecureUrl();
@@ -197,12 +214,14 @@ public class CreateDateOfBirthChangeTests : TestBase
                 .RegisterWith(options);
         });
 
-        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/date-of-birth-changes")
+        var request = new HttpRequestMessage(HttpMethod.Post, "/v3/teachers/name-changes")
         {
             Content = CreateJsonContent(new
             {
                 trn = createPersonResult.Trn,
-                dateOfBirth = newDateOfBirth,
+                firstName = newFirstName,
+                middleName = newMiddleName,
+                lastName = newLastName,
                 evidenceFileName,
                 evidenceFileUrl
             })
